pyproject.toml
[project]
name = "discord-bot-mts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.5.2",
    "faster-whisper>=1.1.1",
    "openai>=1.84.0",
    "py-cord[voice]>=2.6.1",
    "pydantic>=2.11.5",
    "pydub>=0.25.1",
    "pymongo[srv]==3.12",
    "python-dotenv>=1.1.0",
    "supabase>=2.15.2",
]
